+++ README.md
@@ -1,2 +1,2 @@
-# paiya-sd-webui
-这是一个paiya sd webui的库，可用于参考模板生成ai写真。
+# easyphoto-sd-webui
+这是一个easyphoto sd webui的库，可用于参考模板生成ai写真。
